Add some server to config, add fcntl checks to socket new clients

+++ src/ConfigParser.cpp
@@ -173,7 +173,7 @@ void ConfigParser::checkMain(const std::string& keyword, const std::string& valu
 	}
 	else if (keyword.compare(PARSELISTEN) == 0)
 	{
-		size_t port = 0;
+		int port = 0;
 		try
 		{
 			port = std::stoi(value);
@@ -182,6 +182,7 @@ void ConfigParser::checkMain(const std::string& keyword, const std::string& valu
 		{
 			configError("Invalid status code.", lineNumber);
 		}
+
 		if (port > UINT16_MAX || port < 0 || value.empty())
 			configError("Invalid port.", lineNumber);
 		temporaryServer->setListenPort(port);
